backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
groq==0.4.0
transformers==4.36.2
torch==2.1.2
pytest==7.4.3
httpx==0.25.2
